requirements.txt
discord.py
python-dotenv
pymongo
tabula-py
pymupdf